tech(api): move files from scoring BC to evaluation BC
