+++ api/src/certification/evaluation/domain/usecases/score-v3-certification.js
@@ -107,6 +107,7 @@ export const scoreV3Certification = withTransaction(
         algorithm,
         v3CertificationScoring,
         cleaScoringCriteria,
+        scoringDegradationService: services.scoringDegradationService,
       });
 
       const certificationAssessmentHistory = CertificationAssessmentHistory.fromChallengesAndAnswers({
