+++ api/tests/certification/evaluation/unit/domain/services/scoring/scoring-degradation-service_test.js
@@ -1,8 +1,8 @@
-import * as flashAlgorithmService from '../../../../../../src/certification/evaluation/domain/services/algorithm-methods/flash.js';
-import { downgradeCapacity } from '../../../../../../src/certification/scoring/domain/services/scoring-degradation-service.js';
-import { domainBuilder, expect } from '../../../../../test-helper.js';
+import * as flashAlgorithmService from '../../../../../../../src/certification/evaluation/domain/services/algorithm-methods/flash.js';
+import { downgradeCapacity } from '../../../../../../../src/certification/evaluation/domain/services/scoring/scoring-degradation-service.js';
+import { domainBuilder, expect } from '../../../../../../test-helper.js';
 
-describe('Integration | Domain | services | scoringDegradationService', function () {
+describe('Certification | Evaluation | Unit | Domain | services | scoringDegradationService', function () {
   it('should degrade the initial capacity', function () {
     // given
     const initialCapacity = 2;
