+++ api/tests/evaluation/unit/domain/models/CertificationAssessmentScoreV3_test.js
@@ -1,13 +1,13 @@
 import _ from 'lodash';
 
-import { CertificationAssessmentScoreV3 } from '../../../../../../src/certification/scoring/domain/models/CertificationAssessmentScoreV3.js';
-import { ABORT_REASONS } from '../../../../../../src/certification/shared/domain/models/CertificationCourse.js';
-import { config } from '../../../../../../src/shared/config.js';
-import { AnswerStatus } from '../../../../../../src/shared/domain/models/AnswerStatus.js';
-import { status } from '../../../../../../src/shared/domain/models/AssessmentResult.js';
-import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
+import { CertificationAssessmentScoreV3 } from '../../../../../src/certification/evaluation/domain/models/CertificationAssessmentScoreV3.js';
+import { ABORT_REASONS } from '../../../../../src/certification/shared/domain/models/CertificationCourse.js';
+import { config } from '../../../../../src/shared/config.js';
+import { AnswerStatus } from '../../../../../src/shared/domain/models/AnswerStatus.js';
+import { status } from '../../../../../src/shared/domain/models/AssessmentResult.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
-describe('Unit | Domain | Models | CertificationAssessmentScoreV3 ', function () {
+describe('Certification | Evaluation | Unit | Domain | Models | CertificationAssessmentScoreV3 ', function () {
   const maxReachableLevelOnCertificationDate = 7;
 
   const competenceId = 'recCompetenceId';
